Feature/services

+++ Server/FlowHood/src/main/java/org/galactic/flowhood/controller/RequestController.java
@@ -1,13 +1,23 @@
 package org.galactic.flowhood.controller;
 
+import jakarta.validation.Valid;
 import org.galactic.flowhood.domain.dto.request.HouseReqDTO;
+import org.galactic.flowhood.domain.dto.request.RequestReqDTO;
 import org.galactic.flowhood.domain.dto.request.RequestStateReqDTO;
 import org.galactic.flowhood.domain.dto.response.GeneralResponse;
+import org.galactic.flowhood.domain.entities.House;
 import org.galactic.flowhood.domain.entities.Request;
+import org.galactic.flowhood.domain.entities.Role;
+import org.galactic.flowhood.domain.entities.User;
+import org.galactic.flowhood.services.HouseService;
 import org.galactic.flowhood.services.RequestService;
+import org.galactic.flowhood.services.RoleService;
+import org.galactic.flowhood.services.UserService;
+import org.galactic.flowhood.utils.SystemRoles;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,14 +25,25 @@ import java.util.UUID;
 
 @RestController
 @RequestMapping("/api/request")
+@CrossOrigin
 public class RequestController {
     final
     RequestService requestService;
 
-    public RequestController(RequestService requestService) {
+    final UserService userService;
+
+    final RoleService roleService;
+
+    final HouseService houseService;
+
+    public RequestController(RequestService requestService, UserService userService, RoleService roleService, HouseService houseService) {
         this.requestService = requestService;
+        this.userService = userService;
+        this.roleService = roleService;
+        this.houseService = houseService;
     }
 
+    //admin only
     @GetMapping("/")
     public ResponseEntity<GeneralResponse> findAllRequest(){
         try{
@@ -36,6 +57,7 @@ public class RequestController {
         }
     }
 
+    //admin only and authenticated user
     @GetMapping("/{_id}")
     public ResponseEntity<GeneralResponse> findRequestById(@PathVariable("_id") String id){
         try{
@@ -51,19 +73,48 @@ public class RequestController {
         }
     }
 
-    //TODO: think here
+    //Access for anyone excepts visitors
     @PostMapping("/")
-    public ResponseEntity<GeneralResponse> createRequest(@RequestBody HouseReqDTO req){
+    public ResponseEntity<GeneralResponse> createRequest(@RequestBody @Valid RequestReqDTO req, BindingResult errors){
         //validate if role is vigilant or responsible then request must be automatically approved else it should be pending
         try{
-            return GeneralResponse.builder().status(HttpStatus.OK).message("role created").getResponse();
+            if (errors.hasErrors())
+                return GeneralResponse.builder().status(HttpStatus.BAD_REQUEST).message("invalid request").data(errors.getAllErrors()).getResponse();
 
+            House house = houseService.getHouseById(UUID.fromString(req.getHouse()));
+            if(house == null)
+                return GeneralResponse.builder().status(HttpStatus.NOT_FOUND).message("house not found").getResponse();
+
+            User visitor = userService.findUserById(UUID.fromString(req.getVisitor()));
+            if(visitor == null)
+                return GeneralResponse.builder().status(HttpStatus.NOT_FOUND).message("visitor not found").getResponse();
+
+            if(!house.getActive())
+                return GeneralResponse.builder().status(HttpStatus.CONFLICT).message("House can not currently accept visits requests").getResponse();
+
+            User resident = userService.findUserAuthenticated().toEntity();
+
+            //Asign current resident rol to handle petition
+            String residentRol = SystemRoles.RESPONSIBLE.getRole();
+            Role role = roleService.findRoleById(SystemRoles.ADMINISTRATOR.getRole());
+            if(resident.getRoles().contains(role))
+                residentRol = SystemRoles.ADMINISTRATOR.getRole();
+
+            boolean isValidResident = houseService.isResponsibleFromHouse(resident, house);
+            if(isValidResident)
+                residentRol = SystemRoles.RESIDENT.getRole();
+            if(!house.getResponsible().getEmail().equals(resident.getEmail()) && !isValidResident)
+                return GeneralResponse.builder().status(HttpStatus.UNAUTHORIZED).message("You can not create requests for this house").getResponse();
+
+            requestService.createRequestHandler(req, house, resident, visitor, residentRol);
+            return GeneralResponse.builder().status(HttpStatus.OK).message("request created").getResponse();
         }
         catch (Exception e){
             return GeneralResponse.builder().status(HttpStatus.INTERNAL_SERVER_ERROR).getResponse();
         }
     }
 
+    //check if request is from resident or admin
     @DeleteMapping("/{_id}")
     public ResponseEntity<GeneralResponse> deleteRequest(@PathVariable("_id") String id){
         try{
@@ -71,14 +122,28 @@ public class RequestController {
             Request request = requestService.findRequestById(_id);
             if (request == null)
                 return GeneralResponse.builder().status(HttpStatus.NOT_FOUND).message("not found").getResponse();
-            requestService.deleteRequest(request);
-            return GeneralResponse.builder().status(HttpStatus.OK).message("deleted!").getResponse();
+
+            User user = userService.findUserAuthenticated().toEntity();
+
+            Role admRole = roleService.findRoleById(SystemRoles.ADMINISTRATOR.getRole());
+            Role resRole = roleService.findRoleById(SystemRoles.RESPONSIBLE.getRole());
+
+            if(user.getRoles().contains(admRole) ||
+                    (user.getRoles().contains(resRole) && request.getResident()
+                            .getId().equals(user.getId()))){
+                requestService.deleteRequest(request);
+                return GeneralResponse.builder().status(HttpStatus.OK).message("deleted!").getResponse();
+            }
+
+            return GeneralResponse.builder().status(HttpStatus.UNAUTHORIZED).message("could not be deleted!").getResponse();
 
         }
         catch (Exception e){
             return GeneralResponse.builder().status(HttpStatus.INTERNAL_SERVER_ERROR).getResponse();
         }
     }
+
+    //TODO
     @PatchMapping("/{_id}") //getting user validation and status of current request
     public ResponseEntity<GeneralResponse> changeStatus(@PathVariable("_id") String id, @RequestBody RequestStateReqDTO req){
         try{
